pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "beautifulsoup4>=4.14.3",
    "flask>=3.1.2",
    "psycopg2-binary>=2.9.11",
    "requests>=2.32.5",
]
